[ALPS07727967] rrm: Fix coding defect
(1) beacon report might be truncate due to incorrect type casting

MTK-Commit-Id: 32ccc6dd2f58886d1fb0bb52bfe6561bb02cd652

CR-Id: ALPS07727967
Feature: Others
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6898460
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_pf_nep_sanity
CheckSWIP: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>

+++ mgmt/rrm.c
@@ -1506,7 +1506,7 @@ int rrmReportElem(struct RM_MEASURE_REPORT_ENTRY *reportEntry,
 	const uint8_t *data, uint32_t data_len)
 {
 	uint8_t *report = reportEntry->pucMeasReport;
-	uint8_t len = reportEntry->u2MeasReportLen;
+	uint16_t len = reportEntry->u2MeasReportLen;
 	uint32_t size = len + 5 + data_len;
 	uint8_t *buf;
 
@@ -1575,7 +1575,6 @@ int rrmAddBeaconRepElem(struct ADAPTER *prAdapter,
 		goto out;
 
 	pos = buf + ret + sizeof(*rep);
-
 
 	if (!ve) {
 		pos[0] = BEACON_REPORT_SUBELEM_FRAME_BODY_FRAGMENT_ID;
